move RHF folder
