+++ src/components/dialogs/description-modification-dialog.tsx
@@ -12,7 +12,7 @@ import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useSnackMessage } from '../../hooks/useSnackMessage';
 import CustomMuiDialog from './custom-mui-dialog';
-import ExpandingTextField from '../react-hook-form/ExpandingTextField';
+import ExpandingTextField from '../inputs/react-hook-form/ExpandingTextField';
 import { Box } from '@mui/material';
 
 export interface IDescriptionModificationDialog {
