+++ src/components/filter/filter-creation-dialog.tsx
@@ -34,7 +34,7 @@ import {
 import { yupResolver } from '@hookform/resolvers/yup';
 import { elementExistsType } from './criteria-based/criteria-based-filter-edition-dialog';
 import { UUID } from 'crypto';
-import { MergedFormContextProps } from '../react-hook-form/provider/custom-form-provider';
+import { MergedFormContextProps } from '../inputs/react-hook-form/provider/custom-form-provider';
 import { StudyMetadata } from '../../hooks/predefined-properties-hook.ts';
 
 import { FilterContext } from './filter-context';
