+++ src/components/inputs/react-hook-form/directory-items-input.tsx
@@ -13,7 +13,7 @@ import {
     Theme,
     Tooltip,
 } from '@mui/material';
-import OverflowableText from '../OverflowableText';
+import OverflowableText from '../../OverflowableText';
 import FieldLabel from './utils/field-label';
 import FolderIcon from '@mui/icons-material/Folder';
 import {
@@ -27,15 +27,15 @@ import { useController, useFieldArray } from 'react-hook-form';
 import { useIntl } from 'react-intl';
 import MidFormError from '../react-hook-form/error-management/mid-form-error.jsx';
 import { RawReadOnlyInput } from './raw-read-only-input';
-import { mergeSx } from '../../utils/styles.js';
-import DirectoryItemSelector from '../DirectoryItemSelector/directory-item-selector';
+import { mergeSx } from '../../../utils/styles.js';
+import DirectoryItemSelector from '../../DirectoryItemSelector/directory-item-selector';
 import { UUID } from 'crypto';
-import { TreeViewFinderNodeProps } from '../TreeViewFinder/TreeViewFinder';
+import { TreeViewFinderNodeProps } from '../../TreeViewFinder/TreeViewFinder';
 import { useCustomFormContext } from './provider/use-custom-form-context';
 import { isFieldRequired } from './utils/functions';
 import ErrorInput from './error-management/error-input';
-import { useSnackMessage } from '../../hooks/useSnackMessage';
-import { FilterContext } from '../filter/filter-context';
+import { useSnackMessage } from '../../../hooks/useSnackMessage';
+import { FilterContext } from '../../filter/filter-context';
 
 export const NAME = 'name';
