+++ src/components/inputs/react-hook-form/select-inputs/select-input.tsx
@@ -10,7 +10,7 @@ import AutocompleteInput, {
 } from '../autocomplete-inputs/autocomplete-input';
 import { useIntl } from 'react-intl';
 import { FunctionComponent } from 'react';
-import { Option } from '../../../utils/types.ts';
+import { Option } from '../../../../utils/types.ts';
 
 export interface SelectInputProps
     extends Omit<
