+++ src/components/inputs/react-hook-form/slider-input.jsx
