+++ src/components/inputs/react-hook-form/text-input.tsx
@@ -26,7 +26,7 @@ import {
 } from './utils/functions';
 import { useCustomFormContext } from './provider/use-custom-form-context';
 
-import { Input } from '../../utils/types.ts';
+import { Input } from '../../../utils/types.ts';
 
 export interface TextInputProps {
     name: string;
