+++ src/components/react-query-builder-inputs/country-value-editor.tsx
@@ -11,7 +11,7 @@ import { Autocomplete, TextField } from '@mui/material';
 import useConvertValue from './use-convert-value.ts';
 import useValid from './use-valid.ts';
 import { useLocalizedCountries } from '../../hooks/localized-countries-hook';
-import { useCustomFormContext } from '../react-hook-form/provider/use-custom-form-context.ts';
+import { useCustomFormContext } from '../inputs/react-hook-form/provider/use-custom-form-context.ts';
 import { FunctionComponent, useMemo } from 'react';
 
 const CountryValueEditor: FunctionComponent<ValueEditorProps> = (props) => {
