+++ src/components/react-query-builder-inputs/custom-react-query-builder.tsx
@@ -26,8 +26,8 @@ import ValueSelector from './value-selector.tsx';
 import { useCallback, useMemo } from 'react';
 
 import { COMBINATOR_OPTIONS } from '../filter/expert/expert-filter-constants.ts';
-import ErrorInput from '../react-hook-form/error-management/error-input.tsx';
-import FieldErrorAlert from '../react-hook-form/error-management/field-error-alert.tsx';
+import ErrorInput from '../inputs/react-hook-form/error-management/error-input.tsx';
+import FieldErrorAlert from '../inputs/react-hook-form/error-management/field-error-alert.tsx';
 import {
     countRules,
     getOperators,
