+++ src/index.js
@@ -25,9 +25,9 @@ export { default as CustomMuiDialog } from './components/dialogs/custom-mui-dial
 export { default as DescriptionModificationDialog } from './components/dialogs/description-modification-dialog';
 export { default as CriteriaBasedForm } from './components/filter/criteria-based/criteria-based-form';
 export { default as PopupConfirmationDialog } from './components/dialogs/popup-confirmation-dialog';
-export { default as BottomRightButtons } from './components/react-hook-form/ag-grid-table/bottom-right-buttons';
-export { default as CustomAgGridTable } from './components/react-hook-form/ag-grid-table/custom-ag-grid-table';
-export { ROW_DRAGGING_SELECTION_COLUMN_DEF } from './components/react-hook-form/ag-grid-table/custom-ag-grid-table';
+export { default as BottomRightButtons } from './components/inputs/react-hook-form/ag-grid-table/bottom-right-buttons';
+export { default as CustomAgGridTable } from './components/inputs/react-hook-form/ag-grid-table/custom-ag-grid-table';
+export { ROW_DRAGGING_SELECTION_COLUMN_DEF } from './components/inputs/react-hook-form/ag-grid-table/custom-ag-grid-table';
 
 export {
     Line,
@@ -137,25 +137,25 @@ export { useIntlRef } from './hooks/useIntlRef';
 export { useSnackMessage } from './hooks/useSnackMessage';
 export { useDebounce } from './hooks/useDebounce';
 export { default as SelectClearable } from './components/inputs/select-clearable';
-export { useCustomFormContext } from './components/react-hook-form/provider/use-custom-form-context';
-export { default as CustomFormProvider } from './components/react-hook-form/provider/custom-form-provider';
-export { default as AutocompleteInput } from './components/react-hook-form/autocomplete-inputs/autocomplete-input';
-export { default as TextInput } from './components/react-hook-form/text-input';
-export { default as ExpandingTextField } from './components/react-hook-form/ExpandingTextField';
-export { default as RadioInput } from './components/react-hook-form/radio-input';
-export { default as SliderInput } from './components/react-hook-form/slider-input';
-export { default as FloatInput } from './components/react-hook-form/numbers/float-input';
-export { default as IntegerInput } from './components/react-hook-form/numbers/integer-input';
-export { default as SelectInput } from './components/react-hook-form/select-inputs/select-input';
-export { default as CheckboxInput } from './components/react-hook-form/booleans/checkbox-input';
-export { default as SwitchInput } from './components/react-hook-form/booleans/switch-input';
-export { default as ErrorInput } from './components/react-hook-form/error-management/error-input';
-export { default as FieldErrorAlert } from './components/react-hook-form/error-management/field-error-alert';
-export { default as MidFormError } from './components/react-hook-form/error-management/mid-form-error';
-export { default as TextFieldWithAdornment } from './components/react-hook-form/utils/text-field-with-adornment';
-export { default as FieldLabel } from './components/react-hook-form/utils/field-label';
-export { default as SubmitButton } from './components/react-hook-form/utils/submit-button';
-export { default as CancelButton } from './components/react-hook-form/utils/cancel-button';
+export { useCustomFormContext } from './components/inputs/react-hook-form/provider/use-custom-form-context';
+export { default as CustomFormProvider } from './components/inputs/react-hook-form/provider/custom-form-provider';
+export { default as AutocompleteInput } from './components/inputs/react-hook-form/autocomplete-inputs/autocomplete-input';
+export { default as TextInput } from './components/inputs/react-hook-form/text-input';
+export { default as ExpandingTextField } from './components/inputs/react-hook-form/ExpandingTextField';
+export { default as RadioInput } from './components/inputs/react-hook-form/radio-input';
+export { default as SliderInput } from './components/inputs/react-hook-form/slider-input';
+export { default as FloatInput } from './components/inputs/react-hook-form/numbers/float-input';
+export { default as IntegerInput } from './components/inputs/react-hook-form/numbers/integer-input';
+export { default as SelectInput } from './components/inputs/react-hook-form/select-inputs/select-input';
+export { default as CheckboxInput } from './components/inputs/react-hook-form/booleans/checkbox-input';
+export { default as SwitchInput } from './components/inputs/react-hook-form/booleans/switch-input';
+export { default as ErrorInput } from './components/inputs/react-hook-form/error-management/error-input';
+export { default as FieldErrorAlert } from './components/inputs/react-hook-form/error-management/field-error-alert';
+export { default as MidFormError } from './components/inputs/react-hook-form/error-management/mid-form-error';
+export { default as TextFieldWithAdornment } from './components/inputs/react-hook-form/utils/text-field-with-adornment';
+export { default as FieldLabel } from './components/inputs/react-hook-form/utils/field-label';
+export { default as SubmitButton } from './components/inputs/react-hook-form/utils/submit-button';
+export { default as CancelButton } from './components/inputs/react-hook-form/utils/cancel-button';
 export {
     genHelperPreviousValue,
     genHelperError,
@@ -164,16 +164,16 @@ export {
     gridItem,
     isFloatNumber,
     toFloatOrNullValue,
-} from './components/react-hook-form/utils/functions';
+} from './components/inputs/react-hook-form/utils/functions';
 export {
     keyGenerator,
     areArrayElementsUnique,
     mergeSx,
     isObjectEmpty,
 } from './utils/functions';
-export { default as DirectoryItemsInput } from './components/react-hook-form/directory-items-input';
+export { default as DirectoryItemsInput } from './components/inputs/react-hook-form/directory-items-input';
 export { default as DirectoryItemSelector } from './components/DirectoryItemSelector/directory-item-selector';
-export { RawReadOnlyInput } from './components/react-hook-form/raw-read-only-input';
+export { RawReadOnlyInput } from './components/inputs/react-hook-form/raw-read-only-input';
 
 export { default as FilterCreationDialog } from './components/filter/filter-creation-dialog';
 export { default as ExpertFilterEditionDialog } from './components/filter/expert/expert-filter-edition-dialog';
@@ -189,17 +189,17 @@ export {
     DEFAULT_RANGE_VALUE,
     getRangeInputDataForm,
     getRangeInputSchema,
-} from './components/react-hook-form/range-input';
-export { default as InputWithPopupConfirmation } from './components/react-hook-form/select-inputs/input-with-popup-confirmation';
-export { default as MuiSelectInput } from './components/react-hook-form/select-inputs/mui-select-input';
-export { default as CountriesInput } from './components/react-hook-form/select-inputs/countries-input';
+} from './components/inputs/react-hook-form/range-input';
+export { default as InputWithPopupConfirmation } from './components/inputs/react-hook-form/select-inputs/input-with-popup-confirmation';
+export { default as MuiSelectInput } from './components/inputs/react-hook-form/select-inputs/mui-select-input';
+export { default as CountriesInput } from './components/inputs/react-hook-form/select-inputs/countries-input';
 export {
     getSystemLanguage,
     getComputedLanguage,
 } from './hooks/localized-countries-hook';
-export { default as MultipleAutocompleteInput } from './components/react-hook-form/autocomplete-inputs/multiple-autocomplete-input';
-export { default as CsvUploader } from './components/react-hook-form/ag-grid-table/csv-uploader/csv-uploader';
-export { UniqueNameInput } from './components/react-hook-form/unique-name-input';
+export { default as MultipleAutocompleteInput } from './components/inputs/react-hook-form/autocomplete-inputs/multiple-autocomplete-input';
+export { default as CsvUploader } from './components/inputs/react-hook-form/ag-grid-table/csv-uploader/csv-uploader';
+export { UniqueNameInput } from './components/inputs/react-hook-form/unique-name-input';
 export { UserManagerMock } from './utils/UserManagerMock';
 export {
     FILTER_EQUIPMENTS,
